change wordHeight type from int to double for improved precision in word positioning

+++ lib/src/game/config/word_search_game.dart
@@ -8,7 +8,7 @@ class WordSearchConfig {
   final Color primaryColor;
   final Color secondaryColor;
   final double timeLimit;
-  final int wordHeight;
+  final double wordHeight;
 
   WordSearchConfig({
     required this.words,
